requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.0
click>=8.1

# Ingestion
httpx>=0.27
aiokafka>=0.10
aiofiles>=24.1
boto3>=1.35

# Spark
pyspark>=3.5

# Iceberg / Storage
pyiceberg>=0.7
pyarrow>=17.0

# Governance & Quality
great-expectations>=1.0
openlineage-python>=1.18

# Autoscaler / host metrics
psutil>=6.0

# Observability
prometheus-client>=0.21
opentelemetry-api>=1.27
opentelemetry-sdk>=1.27
opentelemetry-exporter-otlp>=1.27
structlog>=24.4

# Testing
pytest>=8.3
pytest-asyncio>=0.24
moto>=5.0
